requirements.txt
bcrypt==3.2.0
boto3==1.16.14
botocore==1.19.14
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
Django==2.2.4
django-environ==0.4.5
django-s3-storage==0.13.4
idna==2.10
ipython-genutils==0.2.0
jmespath==0.10.0
Pillow==8.0.1
pycparser==2.20
Pygments==2.7.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.25.11
uuid==1.30
